requirements.txt
Django==6.0.4
djangorestframework==3.15.1